refactor: avoid overuse of @std/assert

+++ routes/items/[dateISO]/index.tsx
@@ -1,6 +1,4 @@
 import { type Handlers } from "$fresh/server.ts";
-import { assert } from "@std/assert";
-
 import { type State } from "@/lib/context.ts";
 import { deleteItem } from "@/lib/kv/items.ts";
 
@@ -10,7 +8,9 @@ export const handler: Handlers<unknown, State> = {
       return new Response(null, { status: 401 });
     }
 
-    assert(!!ctx.params.dateISO);
+    if ("dateISO" in ctx.params || !!ctx.params.dateISO) {
+      return new Response(null, { status: 404 });
+    }
 
     await deleteItem(ctx.params.dateISO);
